Use enums/booleans for enumerated attributes.
Modify the attribute getters/setters for attributes defined in
the DTD with enumerated values. Use enums for most string values
and boolean where the values in the DTD are "true|false".

The behavior on the getters is currently to return the value of
the member variable, returning null when the value is unset
(false for boolean attributes). This is different from the
generated methods, which returned the default attribute value
as specified in the DTD even when the member variable was unset. The intent
is to model what will be present in the generated XML.

Add tests that verify that the expected enum/boolean is returned
on unmarshal and that the setters result in the correct string in
the XML.