Use enums/booleans for enumerated attributes. (#13)
* Use enums/booleans for enumerated attributes.

Modify the attribute getters/setters for attributes defined in
the DTD with enumerated values. Use enums for most string values
and Boolean where the values in the DTD are "true|false". Remove string setters. Use Enum/Boolean in the member field definitions instead of preserving the original string field
types.

The behavior on the getters is currently to return the value of
the member variable, returning null when the value is unset. This is different from the
generated methods, which returned the default attribute value
as specified in the DTD even when the member variable was unset. The intent
is to model what will be present in the generated XML.

Add tests that verify that the expected enum/Boolean is returned
on unmarshal and that the setters result in the correct string in
the XML. Add tests of fromString with the enum values. Let fromString return
null when passed null. Add string methods to GsafeedHelper and
XmlgroupsHelper. Add tests in RecordTest of invalid attribute values
for enum and boolean attributes when using validation to show the difference in behavior between validation (throws an exception) and no validation (leaves member variable unset).