Use Date for last-modified attribute. (#14)
* Use Date for last-modified attribute.

Change the last-modified attribute on record elements to
marshal/unmarshal as a Date object. The conversion is done using a
custom XmlAdapter, so a ValidationEventHandler has been added when
unmarshalling using the DTD to report parse errors in the adapter.

Use "GMT" instead of "+0000" in dates.

Use java.util.logging for validation messages. Add a JUnit rule to
disable console logging during tests to keep test output readable.